fix: fix bug 6
fix an crucial bug, correct the update of the matrix to be matched. originally the projects
discarded in previous turns wouldn't be removed, and discarded projects was removed one by one,
which affect the index later. Now all the rest projects are appended to the new matrix. The result
is in result8.csv